requirements.txt
streamlit==1.37.0
python-dotenv==1.0.1
google-generativeai==0.5.4
PyPDF2==3.0.1
speechrecognition==3.10.1
pyttsx3==2.90
fpdf==1.7.2
python-docx==1.1.0
protobuf~=5.29.4
dotenv~=0.9.9
